requirements_torch.txt
absl-py==1.4.0
albumentations==1.3.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
blis==0.7.9
brewer2mpl==1.4.1
cachetools==5.3.1
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7
decorator==5.1.1
efficientnet-pytorch==0.7.1
executing==1.2.0
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.39.4
fsspec==2023.6.0
gast==0.4.0
grpcio==1.54.2
h5py==3.8.0
idna==3.4
imageio==2.31.0
ipykernel==6.23.1
ipython==8.14.0
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.2.1
munch==3.0.0
murmurhash==1.0.9
nest-asyncio==1.5.6
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
onnx==1.15.0
onnxruntime==1.16.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.5.1
preshed==3.0.8
pretrainedmodels==0.7.4
prettyplotlib==0.1.7
prompt-toolkit==3.0.38
protobuf==4.23.2
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pydantic==1.10.9
Pygments==2.15.1
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
SemTorch==0.1.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tfimm==0.2.14
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
torch==2.0.1+cu117
torchaudio==2.0.2+cu117
torchvision==0.15.2+cu117
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
wasabi==1.1.2
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.14.1
zipp==3.15.0
